chore: 🤖 Migrate to Jest from AVA
Ava upgraded to 3, removing support for transpilation. Good rationale,
but it meant imports are not allowed anymore. If imports aren't allowed,
then I don't really see a USP to AVA, and therefore want t

BREAKING CHANGE: o consolidate on jest as the test runner I use.
